Update product-plays.html
reorganized the navigation to fit our logical flow that we developed.  Would be great if the design team can come up with a good looking way to have main topics and sub topics.  For example, Project Kickoff is a main topic, and Defining Strategic Objectives is a sub topic of that one.